requirements.txt
torch>=1.9.0
transformers>=4.24.0
fastapi>=0.95.0
uvicorn>=0.21.0
Pillow>=9.0.0
python-dotenv>=1.0.0
boto3>=1.26.0
python-multipart>=0.0.6
h5py>=3.8.0 